audio/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.0
python-multipart==0.0.20
boto3==1.34.30
pydantic==2.9.2
pydantic-settings==2.7.1
websockets==14.2
python-dotenv==1.1.0
